+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Character: add shortcut link to traits from classfeatures tab
+- Character: add shortcut link to magic items from inventory
 
 ## [3.6.2] - 2019-11-16
 ### Fixed